+++ inventory_rest_api/Models/PaymentSales.cs
@@ -7,7 +7,7 @@ namespace inventory_rest_api.Models
         public long PaymentSalesId { get; set; }
         public long SalesId { get; set; }
         public string PaymentSalesDate { get; set; }
-        public long PaymentAmount { get; set; }
+        public double PaymentAmount { get; set; }
 
         [JsonIgnore]
         public Sales Sales { get; set; }
